% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or2001.R
\docType{data}
\name{or2001}
\alias{or2001}
\title{A data set on MID onset; illustrates logistic regression}
\format{
A dyad-year data set.
\describe{
\item{\code{statea}}{the first state in the dyad}
\item{\code{stateb}}{the second state in the dyad}
\item{\code{stateaname}}{the name of first state in the dyad; created with countrycode}
\item{\code{statebname}}{the second state in the dyad; created with countrycode}
\item{\code{year}}{the year}
\item{\code{dispute}}{a factor variable indicating the onset of conflict, specifically the onset of a militarized interstate dispute}
\item{\code{allies}}{a factor variable indicating whether the members of the dyad are allies or not. Allies are linked by a mutual defense treaty, neutrality pact, or entente.}
\item{\code{lcaprat2}}{the logged ratio of the stronger state's military capability index to that of the weaker member}
\item{\code{contiguity}}{factor variable indicating whether the members of the dyad are contiguous or not}
\item{\code{dem.lo}}{the smaller of the two Polity IV scores in the dyad}
\item{\code{logdstab}}{the logged number of miles between the capitals}
\item{\code{power}}{factor variable indicating whether the dyad is minor powers or at least one member is a great power.}
}
}
\usage{
or2001
}
\description{
Oneal and Russett (2001) data set used in Rainey (2016) to illustrate compression in logistic regression models.
These are the data to reproduce the logistic regression coefficients in Table 2 on p. 11 of the \href{https://static.cambridge.org/content/id/urn:cambridge.org:id:article:S204984701500059X/resource/name/S204984701500059Xsup001.pdf}{Appendix to Rainey (2016)}.
Note that I've altered the coding of some variables, so the coefficients won't replicate exactly.
}
\examples{

# a simple example

or <- crdata::or2001

# glm version of their gee on pp. 314 with no product term
m.noprod <- glm(dispute ~ allies + lcaprat2 + contiguity + dem.lo + logdstab + power,
                family = "binomial", data = or)

# glm version of their gee on pp. 314 with no product terms
m.prod <- glm(dispute ~ allies + lcaprat2 + contiguity + dem.lo*logdstab + power,
              family = "binomial", data = or)


# there are no non-contiguous minor powers in this data set
table(or$contiguity, or$power)


}
\keyword{datasets}
